Stop automatic console toggle causing history navigation
If you put a breakpoint at 0xC029 in the training level and press the up
arrow, the console is automatically toggled when the breakpoint gets
hit. However, holding up causes history navigation to occur forcing the
user to clear the input before proceeding. This fix stops that from
happening by forcing the `visible_start_time` of the console to be set
if it isn't currently set during the Up and Down arrow handlers.